Adding a #lookup_or_default method to property classes.
The #lookup method has a side-effect of "building" a new property
class object. E.g. given a user object with property :foo. If the
user object is asked for the value of :foo, a :foo property class
will be built and queried, which is fine.

If that user object is then saved, then we have the side effect of
storing the default value of :foo for that user object.

With the new #lookup_or_default method, we can ask the user object
for it's current value of :foo and safely save it without any side
effects.