Setting the owner instance on new records.
Previously we were correctly referencing the owner instance when
looking up properties. e.g.: `model.properties.lookup(:foo)` would
set `foo.model` correctly.

This works because ActiveRecord's #build does this automatically for
us. `model.properties.build` stores `model_id` on the new property.

However, if model is a new record, ActiveRecord's #build can not
store the `model_id` on the new property because the model does not
have an id yet.

In this case, we're manually setting the model _object_ on the new
property instead of the model_id.