Make property_set idempotent(ish). Only ever call has_many once.
has_many isn't idempotent, and we lose information when we "re-open" a
property_set on the same model w/ the same association name.

This moves all the method creation outside of the has_many and puts it
in the right place via some reflection. This means we get to keep
evaluating the property_set blocks on the same thing, without
destroying information by calling has_many multiple times.