Allow for property tables to live on a separate db
This commit adds `attr_accessor` (to the class level) of
`ActiveRecord::Base` that allows you to set a specific connection class
for the property sets tables.

This is useful if you want to have the tables for your model live on one
database, but all of the property sets tables live on a separate database.

Usage:

```ruby
class Account < MainConnectionClass
  # Ensure you set this _before_ configuring the property sets.
  self.property_sets_connection_class = SeparateDatabase

  property_set :settings do
    property :foo
  end
end
```

Co-authored-by: Thomas Countz <thomascountz@gmail.com>