Test when property tables live on separate db
This commit creates two separate databases `test` and `test_alt_database`
and abstract `ActiveRecord` models (MainDatabase, AltDatabase) that
connect to each of those, respectively. The existing models all inherit
from the `MainDatabase` and thus their behavior is unchanged.

This creates a new model called `Parent::AccountAltDb` which is just a
copy of `Parent::Account` but instead we call
`self.property_sets_connection_class = AltDatabase` on it to instruct
`property_sets` to use the connection (and database) from the
`AltDatabase` class when creating and looking for the property_sets
tables. We also run the same migrations we ran for `Account` but now for
`AccountAltDb` and on the `AltDatabase` connection.

We then run the property_sets specs twice: once with `Parent::Account`
and again with `Parent::AccountAltDb` to ensure that they pass, no
matter which database the property tables live on:

```ruby
describe Parent::Account do
  it_behaves_like "different account models", Parent::Account
end

describe Parent::AccountAltDb do
  it_behaves_like "different account models", Parent::AccountAltDb
end
```